Rename files from cache to metadata_cache
