+++ include/pgduckdb/pgduckdb_metadata_cache.hpp
