+++ src/pgduckdb_metadata_cache.cpp
